Move Cluster into cluster.py. Test fails because of extra reward from epoch in progress is not returned in stake.
